\name{ModelMap-internal}
\alias{diagnostics.function}
\alias{getRasts}
\alias{model.SGB}
\alias{prediction.SGB}
\alias{rF.binary}
\alias{prediction.rF.binary}
\alias{rF.continuous}
\alias{prediction.rF.continuous}
\alias{create.model}
\alias{prediction.model}
\alias{production.prediction}

\title{ Internal ModelMap Functions }
\description{
  Internal Subfunctions
}

\details{
These are not to be called by the user.  
}
\value{

}

\author{ Elizabeth Freeman }

\examples{

}

\keyword{internal}

