\name{ModelMap-internal}

\alias{initialize.device}

\alias{REQUIRE.party}
\alias{REQUIRE.quantregForest}
\alias{REQUIRE.gbm}

\alias{mask.fun}
\alias{mask.graphic}
\alias{explore.continuous}
\alias{explore.categorical}

\alias{CF.list2df}
\alias{CF.int2num}

\alias{varImpPlot.MM}

\alias{check.device.type}
\alias{check.device.type.nodefault}
\alias{check.predList}
\alias{check.model.levels}
\alias{check.model.type}
\alias{check.response.type}
\alias{check.response.name}
\alias{imp.extract.sgb}
\alias{imp.extract.rf}
\alias{imp.extract.qrf}
\alias{imp.extract.cf}
\alias{imp.scale}
\alias{getRasts}

\alias{correlation.function}
\alias{diagnostics.function}

\alias{model.SGB}
\alias{prediction.SGB}

\alias{rF.binary}
\alias{prediction.rF.binary}
\alias{rF.categorical}
\alias{prediction.rF.categorical}
\alias{rF.continuous}
\alias{prediction.rF.continuous}

\alias{CF.binary}
\alias{prediction.CF.binary}
\alias{CF.categorical}
\alias{prediction.CF.categorical}
\alias{CF.continuous}
\alias{prediction.CF.continuous}

\alias{rF.quantreg}
\alias{prediction.rF.quantreg}

\alias{create.model}
\alias{prediction.model}

\alias{projcompare}
\alias{projfix}
\alias{FNcheck}
\alias{grd2gri}
\alias{production.prediction}




\title{ Internal ModelMap Functions }
\description{
  Internal Subfunctions
}

\details{
These are not to be called by the user.  
}


\author{ Elizabeth Freeman }



\keyword{internal}

