\name{ModelMap-internal}
\alias{varImpPlot.MM}
\alias{check.model.type}
\alias{imp.extract.sgb}
\alias{imp.extract.rf}
\alias{imp.scale}
\alias{getRasts}
\alias{diagnostics.function}
\alias{model.SGB}
\alias{prediction.SGB}
\alias{rF.binary}
\alias{prediction.rF.binary}
\alias{rF.categorical}
\alias{prediction.rF.categorical}
\alias{rF.continuous}
\alias{prediction.rF.continuous}
\alias{create.model}
\alias{prediction.model}
\alias{production.prediction}


\title{ Internal ModelMap Functions }
\description{
  Internal Subfunctions
}

\details{
These are not to be called by the user.  
}


\author{ Elizabeth Freeman }



\keyword{internal}

